Add CMake build system and updated README.md
