+++ .gitignore
@@ -4,12 +4,43 @@
 *.dll
 *.def
 *.exe
-/lib/slang*
-/mapslang/pkg/*
+*.so
+*.a
+*.pc
+*.log
+lib/slang*
+mapslang/pkg/*
+mapslang/mapslang
+mapslang/map
+map
 .vs/
 .git
 /.cproject
 /.project
 /.settings/
-/tmp/
-/build/
+tmp/
+build/
+Makefile
+CMakeFiles/
+CMakeCache.txt
+cmake_install.cmake
+install_manifest.txt
+libadikted/CMakeFiles/
+libadikted/CMakeCache.txt
+libadikted/Makefile*
+libadikted/cmake_install.cmake
+libadikted/install_manifest.txt
+mapslang/CMakeFiles/
+mapslang/CMakeCache.txt
+mapslang/cmake_install.cmake
+mapslang/Makefile*
+mapslang/install_manifest.txt
+mapslang/map
+putemple
+puttrain
+putgems
+viewmap*
+./putemple
+./puttrain
+./putgems
+./viewmap*
