+++ examples/putemple/putemple_private.h
