+++ examples/viewmap/Makefile.win
