+++ examples/viewmap/viewmap_private.h
