+++ examples/viewmap/viewmap_private.rc
