+++ libadikted/Makefile.win
