+++ libadikted/libadikted.pc.in
@@ -1,0 +1,12 @@
+prefix=@install_prefix@
+exe_prefix=${prefix}
+libdir=${prefix}/@install_libdir@
+includedir=${prefix}/@install_includedir@
+
+Name: lib@target_name@
+Description: @PROJECT_DESCRIPTION@
+Version: @PROJECT_VERSION@
+
+Requires:
+Libs: -L${libdir} -l@target_name@ @stdclib@
+Cflags: -I${includedir} -I${includedir}/@target_name@
