+++ mapslang/Makefile.win
@@ -82,15 +82,15 @@ WARNFLAGS          = -Wall -Wextra -Werror-implicit-function-declaration -Wno-co
 PLATFORM_FLAGS     = -march=i486
 CSTANDARD          = -std=gnu99
 CC                 = $(CROSS_COMPILE)gcc$(CROSS_EXEEXT) $(CSTANDARD) $(WARNFLAGS)
-CFLAGS             = -c -O -I. -I../lib/$(SLANG_NAME)/src $(DLL_FLAG) $(PLATFORM_FLAGS)
-CFLAGS_DEBUG       = -c -g -I. -I../lib/$(SLANG_NAME)/src $(DLL_FLAG) $(PLATFORM_FLAGS) -DDEBUG
+CFLAGS             = -c -O -I.. -I. -I../lib/$(SLANG_NAME)/src $(DLL_FLAG) $(PLATFORM_FLAGS)
+CFLAGS_DEBUG       = -c -g -I.. -I. -I../lib/$(SLANG_NAME)/src $(DLL_FLAG) $(PLATFORM_FLAGS) -DDEBUG
 LINK               = $(CROSS_COMPILE)gcc$(CROSS_EXEEXT)
 WINDRES            = $(CROSS_COMPILE)windres$(CROSS_EXEEXT)
 RES                = obj/map_private.res
 RC                 = map_private.rc
 LFLAGS             = -static-libgcc $(PLATFORM_FLAGS)
 LFLAGS_DEBUG       = -static-libgcc $(PLATFORM_FLAGS) -DDEBUG
-LIBS               = -L. -L../libadikted/bin -L..$(SLANG_BIN_FOLDER) -lslang -ladikted
+LIBS               = -L.. -L. -Llibadikted/bin -L..$(SLANG_BIN_FOLDER) -lslang -ladikted
 SRC                = \
 input_kb.c \
 main.c \
