+++ mapslang/scr_actn.c
@@ -23,7 +23,7 @@
 
 #include <math.h>
 #include <string.h>
-#include "../libadikted/adikted.h"
+#include "libadikted/adikted.h"
 #include "var_utils.h"
 #include "output_scr.h"
 #include "input_kb.h"
