+++ mapslang/scr_clm.c
@@ -21,7 +21,7 @@
 
 #include "scr_clm.h"
 
-#include "../libadikted/adikted.h"
+#include "libadikted/adikted.h"
 #include "output_scr.h"
 #include "input_kb.h"
 #include "scr_actn.h"
