+++ mapslang/scr_slab.c
@@ -22,7 +22,7 @@
 #include "scr_slab.h"
 
 #include <time.h>
-#include "../libadikted/adikted.h"
+#include "libadikted/adikted.h"
 #include "output_scr.h"
 #include "input_kb.h"
 #include "scr_actn.h"
