+++ mapslang/scr_txted.c
@@ -20,7 +20,7 @@
 
 #include "scr_txted.h"
 
-#include "../libadikted/adikted.h"
+#include "libadikted/adikted.h"
 #include "output_scr.h"
 #include "input_kb.h"
 #include "scr_actn.h"
