+++ mapslang/textmenu.c
@@ -25,10 +25,11 @@
 
 #include "textmenu.h"
 
-#include "../libadikted/globals.h"
+#include "libadikted/globals.h"
 #include "input_kb.h"
 #include "output_scr.h"
 #include "var_utils.h"
+
 
 inline union _value menu_value_num(long num)
 { union _value ret; ret.num=num; return ret; }
