+++ mapslang/var_utils.c
@@ -19,11 +19,11 @@
 
 #include "var_utils.h"
 
-#include "../libadikted/globals.h"
+#include "libadikted/globals.h"
 #include <stdarg.h>
 #include "input_kb.h"
-#include "../libadikted/msg_log.h"
-#include "../libadikted/lev_data.h"
+#include "libadikted/msg_log.h"
+#include "libadikted/lev_data.h"
 #include "output_scr.h"
 #include "scr_actn.h"
